#2033 Split access_info into two new classes.
