+++ doc/developer_guide/dependency.rst
@@ -275,7 +275,7 @@ The only exception to this is if a kernel is called, in which case the
 metadata for the kernel declaration will be used to determine the variable
 accesses for the call statement. The information about all variable usage
 of a PSyIR node or a list of nodes can be gathered by creating an object of
-type `psyclone.core.access_info.VariablesAccessInfo`.
+type `psyclone.core.variables_access_info.VariablesAccessInfo`.
 This class uses a `Signature` object to keep track of the variables used.
 
 Signature
@@ -289,7 +289,7 @@ three components `a`, `b`, and `c`.
 A simple variable such as `a` is stored as a one-element tuple `(a, )`, having
 a single component.
 
-.. autoclass:: psyclone.core.access_info.Signature
+.. autoclass:: psyclone.core.Signature
     :members:
     :special-members: __hash__, __eq__, __lt__
 
@@ -305,7 +305,7 @@ of `VariablesAccessInfo`.
 
 .. automethod:: psyclone.psyir.nodes.Node.reference_accesses
 
-.. autoclass:: psyclone.core.access_info.VariablesAccessInfo
+.. autoclass:: psyclone.core.VariablesAccessInfo
     :members:
     :special-members: __str__
 
@@ -342,24 +342,24 @@ PSyIR operators above will be reported as read access.
 SingleVariableAccessInfo
 ------------------------
 The class `VariablesAccessInfo` uses a dictionary of
-`psyclone.core.access_info.SingleVariableAccessInfo` instances to map
+`psyclone.core.SingleVariableAccessInfo` instances to map
 from each variable to the accesses of that variable. When a new variable
 is detected when adding access information to a `VariablesAccessInfo` instance
 via `add_access()`, a new instance of `SingleVariableAccessInfo` is added,
 which in turn stores all access to the specified variable.
 
-.. autoclass:: psyclone.core.access_info.SingleVariableAccessInfo
+.. autoclass:: psyclone.core.SingleVariableAccessInfo
     :members:
 
 AccessInfo
 ----------
 The class `SingleVariableAccessInfo` uses a list of
-`psyclone.core.access_info.AccessInfo` instances to store all
+`psyclone.core.AccessInfo` instances to store all
 accesses to a single variable. A new instance of `AccessInfo`
 is appended to the list whenever `add_access_with_location()`
 is called.
 
-.. autoclass:: psyclone.core.access_info.AccessInfo
+.. autoclass:: psyclone.core.single_variable_access_info.AccessInfo
     :members:
 
 Indices
@@ -371,7 +371,7 @@ to analyse a PSyIR tree for details. The indices are stored in the
 ComponentIndices object that each access has, which can be accessed
 using the `component_indices` property of an `AccessInfo` object.
 
-.. autoclass:: psyclone.core.access_info.ComponentIndices
+.. autoclass:: psyclone.core.ComponentIndices
     :members:
     :special-members: __getitem__, __len__
